// frontend/src/api/models/chatMessageResponseDtoDTO.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API
 * Clean Architecture API with JWT Authentication
 * OpenAPI spec version: v1
 */

export interface ChatMessageResponseDtoDTO {
  id?: string;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  content?: string | null;
  order?: number;
  createdAt?: string;
}